cleanup remaster #1
