+++ Makefile
@@ -1,6 +1,6 @@
 # Compiler flags
 DEBUG_FLAGS := -DDEBUG -DLOG_FILE=\"/tmp/fensterchef-log.txt\" -g -fsanitize=address
-C_FLAGS := -Wall -Wextra -Wpedantic -Werror -Wno-format-zero-length
+C_FLAGS := -Iinclude -Wall -Wextra -Wpedantic -Werror -Wno-format-zero-length
 RELEASE_FLAGS := -O3
 
 # Libs
