+++ include/fensterchef.h
@@ -13,6 +13,7 @@
 
 #include "action.h"
 #include "frame.h"
+#include "window.h"
 
 /* flag used to configure window position, size and border width */
 #define XCB_CONFIG_SIZE (XCB_CONFIG_WINDOW_X | \
@@ -72,11 +73,6 @@ int init_fensterchef(void);
 /* Quit the window manager and clean up resources. */
 void quit_fensterchef(int exit_code);
 
-/* Shows a windows list where the user can select a window from.
- * -- Implemented in window_list.c --
- */
-Window *select_window_from_list(void);
-
 /* Show the notification window with given message at given coordinates for
  * NOTIFICATION_DURATION seconds.
  *
@@ -84,65 +80,5 @@ Window *select_window_from_list(void);
  * @y Center y position.
  */
 void set_notification(const FcChar8 *msg, int32_t x, int32_t y);
-
-/* -- Implemented in event.c -- */
-
-/* Handle the given xcb event. */
-void handle_event(xcb_generic_event_t *event);
-
-/* -- Implemented in render_font.c -- */
-
-/* Creates a pixmap with width and height set to 1. */
-xcb_pixmap_t create_pen(xcb_render_color_t color);
-
-/* Initializes all parts needed for drawing fonts.
- *
- * @return 0 on success, 1 otherwise.
- */
-int init_font_drawing(void);
-
-/* Frees all resources associated to font rendering. */
-void deinit_font_drawing(void);
-
-/* This sets the globally used font for rendering.
- *
- * @name can be NULL to set the default font.
- *
- * @return 0 on success or 1 when the font was not found.
- */
-int set_font(const FcChar8 *query);
-
-/* This draws text to a given drawable using the current font. */
-void draw_text(xcb_drawable_t xcb_drawable, const FcChar8 *utf8, uint32_t len,
-        xcb_render_color_t background_color, xcb_rectangle_t *rect,
-        xcb_render_picture_t foreground, int32_t x, int32_t y);
-
-/* Measure a text that has no new lines. */
-struct text_measure {
-    int32_t ascent;
-    int32_t descent;
-    uint32_t total_width;
-};
-
-void measure_text(const FcChar8 *utf8, uint32_t len, struct text_measure *tm);
-
-/* -- Implemented in keyboard.c -- */
-
-/* Get a keysym from a keycode. */
-xcb_keysym_t get_keysym(xcb_keycode_t keycode);
-
-/* Get a list of keycodes from a keysym.
- * The caller needs to free this list.
- */
-xcb_keycode_t *get_keycodes(xcb_keysym_t keysym);
-
-/* Grab all keys so we receive the keypress events for them. */
-int setup_keys(void);
-
-/* Get an action code from a key press event (keybind).
- *
- * @return the action index or ACTION_NULL if the bind does not exist.
- */
-action_t get_action_bind(xcb_key_press_event_t *event);
 
 #endif
