+++ include/window_list.h
@@ -1,0 +1,7 @@
+#ifndef WINDOW_LIST_H
+#define WINDOW_LIST_H
+
+/* Shows a windows list where the user can select a window from. */
+Window *select_window_from_list(void);
+
+#endif
